Make internal Request an enum
Also drop tuple conversions; preparing for a Batch variant.

+++ netlink-proto/src/protocol/protocol.rs
@@ -131,7 +131,7 @@ where
     }
 
     pub fn request(&mut self, request: Request<T, M>) {
-        let Request {
+        let Request::Single {
             mut message,
             metadata,
             destination,
